+++ frontend/src/app/page.tsx
@@ -6,6 +6,7 @@ import Link from 'next/link';
 import { User, MessageCircle, ClipboardCheck, ShoppingBag, HomeIcon, Menu, X, Sparkles, FileText } from "lucide-react";
 import { usePathname, useRouter } from 'next/navigation';
 import { useCurrentUser } from '@/data/useCurrentUser';
+import { mbtiList } from '@/data/mbtiList';
 import { apiConfig } from '@/config/api';
 
 // 서버가 내려주는 타입 (영문 키)
@@ -24,84 +25,6 @@ const products = [
   { name: "진정 세럼", description: "피부 진정케어 세럼 민감 피부용", category: "진정" },
   { name: "보습 크림", description: "저자극 수분 크림 민감 피부용", category: "보습" },
 ];
-
-export const mbtiList = {
-  MBST: {
-    type: "민감성",
-    description: "피부가 건조해 당김이 느껴지고, 유분 분비가 안정적이어서 번들거림이 적으며, 탄력은 좋지만 자극에 민감한 상태입니다.",
-    advice: "고보습 크림과 오일로 유수분 균형을 맞추고 가벼운 수분젤로 촉촉함을 유지하면서 저자극·진정 포뮬러로 장벽을 강화하고 콜라겐·펩타이드 성분으로 탄력을 관리하세요."
-  },
-  MBSL: {
-    type: "민감성",
-    description: "피부가 건조해 당김이 느껴지고, 유분 분비가 안정적이어서 번들거림이 적으며, 자극에 민감하고 처짐이 느껴져 탄력 저하가 있습니다.",
-    advice: "고보습 크림과 오일로 유수분 균형을 맞추고 가벼운 수분젤로 촉촉함을 유지하며, 저자극·진정 포뮬러로 장벽을 강화하고 리프팅 세럼과 마사지를 병행해 탄력을 높이세요."
-  },
-  MOIT: {
-    type: "지성",
-    description: "수분과 유분이 모두 충분해 번들거림이 적고, 자극에도 안정적이며, 탄력도 좋은 균형 상태입니다.",
-    advice: "히알루론산 세럼으로 수분 장벽을 강화하고 가벼운 수분젤로 촉촉함을 유지하며, 기본 보습·톤업 케어를 꾸준히 하면서 콜라겐·펩타이드 제품으로 탄력을 유지하세요."
-  },
-  MOIL: {
-    type: "지성",
-    description: "수분과 유분이 충분해 번들거림이 적지만, 탄력은 다소 떨어져 처짐이 느껴집니다.",
-    advice: "히알루론산 세럼으로 수분 장벽을 강화하고 가벼운 수분젤로 촉촉함을 유지하면서, 기본 보습·톤업 케어를 하되 리프팅 세럼과 마사지를 병행해 탄력을 높이세요."
-  },
-  MBIT: {
-    type: "복합성",
-    description: "수분은 충분하나 유분이 부족해 거칠어지고, 자극에 민감하면서도 탄력은 좋은 상태입니다.",
-    advice: "히알루론산 세럼으로 수분을 보충하고 리치 오일·크림으로 유분을 채우며, 저자극·진정 포뮬러로 장벽을 강화하고 콜라겐·펩타이드 성분으로 탄력을 관리하세요."
-  },
-  MBIL: {
-    type: "복합성",
-    description: "수분은 충분하지만 유분과 탄력이 모두 부족해 당김과 처짐이 함께 나타납니다.",
-    advice: "히알루론산 세럼으로 수분을 보충하고 리치 오일·크림으로 유분을 채운 뒤, 저자극·진정 포뮬러로 장벽을 강화하고 리프팅 세럼·마사지를 통해 탄력을 높이세요."
-  },
-  DOST: {
-    type: "수분부족지성",
-    description: "수분이 부족해 당김이 느껴지지만 유분과 탄력은 안정적인 상태입니다.",
-    advice: "고보습 크림과 오일로 수분을 채우고, 저자극 수분 세럼과 마스크로 수분 장벽을 강화하면서 콜라겐·펩타이드 성분으로 탄력을 유지하세요."
-  },
-  DOSL: {
-    type: "수분부족지성",
-    description: "수분과 탄력이 부족해 당김과 처짐이 동시에 느껴집니다.",
-    advice: "고보습 크림과 오일로 수분을 채우고 저자극 진정 세럼으로 장벽을 강화하며 리프팅 세럼·마사지로 탄력을 개선하세요."
-  },
-  DBST: {
-    type: "건성",
-    description: "수분과 유분이 모두 부족해 건조함이 심하지만 탄력은 비교적 좋은 상태입니다.",
-    advice: "고보습 크림과 오일로 집중 영양을 공급하고, 기본 보습·톤업 케어를 유지하며 콜라겐·펩타이드 성분으로 탄력을 관리하세요."
-  },
-  DBSL: {
-    type: "건성",
-    description: "수분·유분·탄력이 모두 부족해 피부가 거칠고 처짐이 심합니다.",
-    advice: "장벽 강화·고보습 크림과 오일로 영양을 채우고, 리프팅 세럼·마사지로 탄력을 강화하세요."
-  },
-  DOIT: {
-    type: "수분부족지성",
-    description: "수분이 부족해 당김이 느껴지지만 유분과 탄력은 좋은 상태입니다.",
-    advice: "수분 에센스와 마스크로 즉각적인 수분을 보충하고, 콜라겐·펩타이드 성분으로 탄력을 유지하세요."
-  },
-  DOIL: {
-    type: "수분부족지성",
-    description: "유분은 충분하지만 수분과 탄력 모두 부족해 당김과 처짐이 느껴집니다.",
-    advice: "고보습 세럼으로 수분을 보충하고 탄력 강화 오일을 함께 사용해 탄력을 회복하세요."
-  },
-  DBIT: {
-    type: "건성",
-    description: "수분·유분이 부족해 건조함이 느껴지지만 탄력은 비교적 유지되는 상태입니다.",
-    advice: "보습 크림과 수분 세럼으로 영양을 보충하고 콜라겐·펩타이드 성분으로 탄력을 관리하세요."
-  },
-  DBIL: {
-    type: "건성",
-    description: "수분·유분·탄력이 모두 부족해 건조하고 처짐이 심한 상태입니다.",
-    advice: "장벽 강화·고보습 크림과 탄력 세럼을 병행해 집중 관리하세요."
-  },
-  default: {
-    type: "표준형",
-    description: "피부 균형이 잘 잡힌 상태입니다.",
-    advice: "기본 보습과 탄력 관리 루틴을 꾸준히 지켜주세요."
-  }
-};
 
 
 export default function Home() {
